+++ packages/const/src/index.ts
@@ -1,5 +1,6 @@
+export * from './branding';
 export * from './image';
-export * from './locale';
+export * from './layoutTokens';
 export * from './message';
 export * from './settings';
 export * from './version';