+++ src/app/[variants]/(main)/files/[id]/page.tsx
@@ -1,3 +1,4 @@
+import { getUserAuth } from '@lobechat/utils/server';
 import { notFound } from 'next/navigation';
 import { Flexbox } from 'react-layout-kit';
 
@@ -5,7 +6,6 @@ import FileViewer from '@/features/FileViewer';
 import { createCallerFactory } from '@/libs/trpc/lambda';
 import { lambdaRouter } from '@/server/routers/lambda';
 import { PagePropsWithId } from '@/types/next';
-import { getUserAuth } from '@/utils/server/auth';
 
 import FileDetail from '../features/FileDetail';
 import Header from './Header';